balar-mmio: update test binary
